requirements.txt
numpy==1.26.3
pandas==2.2.3
ultralytics==8.3.51
opencv-python==4.10.0.84
torch==2.5.1
PyYAML==6.0.2
pydantic==2.10.4 
